+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.6.6
 * KeyVault: Enable VaultUri configuration member for use as output parameter.
+* KeyVault: Fix emitted `enablePurgeProtection`.
 
 ## 1.6.5
 * Azure Firewall: Bug fix for link_to_vhub and added depends_on to builder